pyproject.toml
[tool.poetry]
name = "banking-analysis"
version = "0.1.0"
description = "AI-powered banking analysis tool using CrewAI"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = ">=2.4.2,<3.0.0"
python-dotenv = "^1.0.0"
crewai = "^0.1.0"
pandas = "^2.2.0"
psycopg2-binary = "^2.9.9"
python-jose = "^3.3.0"
requests = "^2.31.0"
gspread = "^5.12.0"
google-auth = "^2.27.0"
fpdf2 = "^2.7.8"
langchain = "^0.1.0"
aiohttp = "^3.9.3"
PyJWT = "^2.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
black = "^24.1.0"
isort = "^5.13.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 